pyproject.toml
[project]
name = "mcp-server-selenium"
version = "0.1.0"
description = "A Model Context Protocol server providing tools to access a chrome browser through Selenium"
authors = [
    { name = "Daniel de Cordoba", email = "danidecordoba@gmail.com" },
]
keywords = ["selenium", "mcp", "llm"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fake-useragent>=2.2.0",
    "mcp[cli]>=1.6.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.76.2",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "selenium>=4.31.0",
    "selenium-stealth>=1.0.6",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
]

[project.scripts]
mcp-server-selenium = "mcp_server_selenium:main"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipython>=9.2.0",
    "isort>=6.0.1",
    "pylint>=3.3.6",
    "ruff>=0.11.7",
]

[tool.black]
line-length = 88
target_version = ['py311']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

[tool.isort]
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0

[tool.pylint]
max-line-length = "88"
disable=[
    "C0114", # Missing-module-docstring
    "C0115", # Missing-class-docstring
    "C0116", # Missing function or method docstring
    "R0901", # Too many ancestors (N/7)
    "R0903", # Too few public methods (N/2)
    "R0913", # Too many arguments (N/5)
    "R0914", # Too many local variables (N/15)
    "R0801", # Similar lines in N files
]

[tool.ruff]
target-version = "py311"
  [tool.ruff.lint]
  select = [
      "E",  # pycodestyle errors
      "W",  # pycodestyle warnings
      "F",  # pyflakes
      "I",  # isort
      "B",  # flake8-bugbear
      "C4",  # flake8-comprehensions
      "UP",  # pyupgrade
  ]
